+++ src/main/java/com/jobtracker/job_tracker/service/UserService.java
@@ -59,6 +59,9 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        if(!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
